Проверка ДЗ | 1. Начало работы
- Правильно: с 0 по 9
- Неправильно:
- Улучшены:
- На доработку:

+++ tasks/task_8.py
@@ -1,6 +1,6 @@
 a = "3.8"
 b = "15"
 
-multi = float(a)*float(b)  # здесь Ваш код
+multi = int(float(a)) * int(b)  # здесь Ваш код
 
 print("Произведение чисел:", multi)  # допишите код
